Fixed path search method for node_modules
