+++ llrt_core/Cargo.toml
@@ -79,6 +79,8 @@ flate2 = { version = "1.0.30", features = [
 ], default-features = false }
 brotlic = "0.8.2"
 rustls-pemfile = "2.1.2"
+lazy_static = "1.5.0"
+home = "0.5.9"
 
 [build-dependencies]
 rquickjs = { version = "0.6.2", features = [
