+++ llrt_core/src/lib.rs
@@ -13,6 +13,7 @@ mod compiler_common;
 pub mod environment;
 pub mod json;
 // mod minimal_tracer;
+mod custom_resolver;
 mod module_builder;
 pub mod modules;
 pub mod number;
